Find backends within rules.
Before this change, a global CLOUD_PROVIDER option was required to
tell us which backed to check for existing snapshots.

Now, we look at all the rules, collect the backends in use, and load
the snapshots from all of them.

- No configuration required.
- Multiple disk-backends on the same cloud supported.

Note that the second part may not actually be functional. There may
be some issues there regarding matching snapshots to the backends
during expiration. But the architecture in principle would allow this
case.