Re-open k8s watches that close successfully
When a watch finishes (without errors), this indicates the API server
closed the connection, eg. because it's shutting down.

We do not want to stop watching in this case, so we re-open a new
connection and continue watching. This will result in re-processing
everything, which is fine.

This prevents a condition where the watch returns, resulting in a state
where the process no longer responds to any kubernetes changes,
preventing new snapshot rules from being picked up.