Allow SnapshotRule to refer to a volume claim.
This also refactors some of the "rule finding code". I think it's
easier to understand now. One thing that changed is that we no
longer try to follow a "volume" resource via "claimRef" to the volume
claim, to search for deltas there. We also watched claim resources
independantly of volumes, which should be enough to handle them.