Discard scheduled snapshots if they are no longer required.
If there are unrelated PV- or PVC-events in the cluster while a scheduled snapshot has not completed the snapshot will be scheduled and processed again.
Since each snapshot will trigger another schedule this leads to the queue filling up endlessly.
To prevent the queue from filling because of these duplicate backups we added a check if the backup is still required before starting it and will discard the snapshot if it isn't necessary anymore.